Fixed find_pos_of_max_double bug

+++ statisticsseq.c
@@ -30,7 +30,7 @@ int find_max(int* grades, int nStudents){
 
 // Calculate the position of the maximum value of an array
 int find_pos_of_max_double(double* vet, int size){
-	int max = vet[0];
+	double max = vet[0];
 	int pos = 0;
 	int i;
 	for(i=1; i<size; i++){
